Support study/answer activity

+++ buddypress/activity/entry.php
@@ -6,7 +6,7 @@
  * This template is used by activity-loop.php and AJAX functions to show
  * each activity.
  *
- * @package BuddyPress
+ * @package    BuddyPress
  * @subpackage bp-legacy
  */
 
@@ -16,20 +16,37 @@ global $activities_template;
 <?php do_action( 'bp_before_activity_entry' ); ?>
 
 <li class="<?php bp_activity_css_class(); ?>" id="activity-<?php bp_activity_id(); ?>">
-	<div class="activity-avatar">
-		<?php bp_activity_avatar(); ?>
-	</div>
 
 	<div class="activity-content">
 
 		<div class="activity-header">
-			<p class="small">
-				<?php echo $activities_template->activity->action; ?> <br />
-				<span class="time-since"><?php echo bp_core_time_since( $activities_template->activity->date_recorded ); ?></span>
-			</p>
+			<div class="activity-avatar">
+				<?php bp_activity_avatar(); ?>
+			</div>
+
+			<?php if ( 'sc_study' == get_post_type() ) : ?>
+				<p class="small">
+					<span class="time-since"><?php echo bp_core_time_since( $activities_template->activity->date_recorded ); ?></span><br />
+					<?php if ( get_current_user_id() == bp_get_activity_user_id() ) : ?>
+						<?php _e( 'Your answer', 'sc' ); ?>
+					<?php else : ?>
+						<?php echo bp_core_get_user_displayname( bp_get_activity_user_id() ); ?> <?php _e( 'answered', 'sc' ); ?>
+					<?php endif; ?>
+				</p>
+			<?php else : ?>
+				<p class="small">
+					<?php echo $activities_template->activity->action; ?>
+					<?php if ( is_page_template( 'templates/profile.php' ) && $group = groups_get_group( 'group_id=' . bp_get_activity_item_id() ) ) : ?>
+						&nbsp;|&nbsp;<a href="<?php bp_group_permalink( $group ); ?>"><?php bp_group_name( $group ); ?></a>
+					<?php endif; ?>
+					<br />
+					<span class="time-since"><?php echo bp_core_time_since( $activities_template->activity->date_recorded ); ?></span>
+				</p>
+			<?php endif; ?>
 		</div>
 
-		<?php if ( bp_activity_has_content() ) : ?>
+		<?php // only show answer update content when there is a discussion on it. ?>
+		<?php if ( bp_activity_has_content() && ( 'sc_study' == get_post_type() || 'answer_update' != bp_get_activity_type() || bp_activity_get_comment_count() ) ) : ?>
 
 			<div class="activity-inner">
 
@@ -41,31 +58,37 @@ global $activities_template;
 
 		<?php do_action( 'bp_activity_entry_content' ); ?>
 
-		<p class="activity-meta small">
+		<?php if ( 'answer_update' != bp_get_activity_type() || bp_activity_get_comment_count() || 'sc_study' == get_post_type() ) : ?>
+			<p class="activity-meta small">
 
-			<?php if ( bp_get_activity_type() == 'activity_comment' ) : ?>
+				<?php if ( bp_get_activity_type() == 'activity_comment' ) : ?>
 
-				<a href="<?php bp_activity_thread_permalink(); ?>" class="view" title="<?php esc_attr_e( 'View Conversation', 'buddypress' ); ?>"><?php _e( 'View Conversation', 'buddypress' ); ?></a>&nbsp;|&nbsp;
-
-			<?php endif; ?>
-
-			<?php if ( is_user_logged_in() ) : ?>
-
-				<?php if ( bp_activity_can_comment() ) : ?>
-
-					<a href="<?php bp_activity_comment_link(); ?>" class="acomment-reply" id="acomment-comment-<?php bp_activity_id(); ?>"><?php _e( 'Add your comment', 'sc' ) ?></a>
+					<a href="<?php bp_activity_thread_permalink(); ?>" class="view" title="<?php esc_attr_e( 'View Conversation', 'buddypress' ); ?>"><?php _e( 'View Conversation', 'buddypress' ); ?></a>&nbsp;|&nbsp;
 
 				<?php endif; ?>
 
-				<?php if ( bp_activity_user_can_delete() ) : ?>
-					&nbsp;|&nbsp;<a href="<?php bp_activity_delete_url(); ?>"><?php _e( 'Delete', 'sc' ); ?></a>
+				<?php if ( is_user_logged_in() ) : ?>
+
+					<?php if ( bp_activity_get_comment_count() ) : ?>
+						<a href="#" class="toggle-comments" id="toggle-comments-<?php bp_activity_id(); ?>"><?php printf( __( 'Hide Replies (%s)', 'sc' ), bp_activity_get_comment_count() ); ?></a>&nbsp;|&nbsp;
+					<?php endif; ?>
+
+					<?php if ( bp_activity_can_comment() ) : ?>
+
+						<a href="<?php bp_activity_comment_link(); ?>" class="acomment-reply" id="acomment-comment-<?php bp_activity_id(); ?>"><?php _e( 'Add your comment', 'sc' ) ?></a>
+
+					<?php endif; ?>
+
+					<?php if ( bp_activity_user_can_delete() ) : ?>
+						&nbsp;|&nbsp;<a href="<?php bp_activity_delete_url(); ?>"><?php _e( 'Delete', 'sc' ); ?></a>
+					<?php endif; ?>
+
+					<?php do_action( 'bp_activity_entry_meta' ); ?>
+
 				<?php endif; ?>
 
-				<?php do_action( 'bp_activity_entry_meta' ); ?>
-
-			<?php endif; ?>
-
-		</p>
+			</p>
+		<?php endif; ?>
 
 	</div>
 
@@ -85,7 +108,8 @@ global $activities_template;
 						<div class="ac-textarea">
 							<textarea id="ac-input-<?php bp_activity_id(); ?>" class="ac-input bp-suggestions" name="ac_input_<?php bp_activity_id(); ?>"></textarea>
 						</div>
-						<input type="submit" name="ac_form_submit" value="<?php esc_attr_e( 'Post', 'buddypress' ); ?>" /> &nbsp; <a href="#" class="ac-reply-cancel"><?php _e( 'Cancel', 'buddypress' ); ?></a>
+						<input type="submit" name="ac_form_submit" value="<?php esc_attr_e( 'Post', 'buddypress' ); ?>" /> &nbsp;
+						<a href="#" class="ac-reply-cancel"><?php _e( 'Cancel', 'buddypress' ); ?></a>
 						<input type="hidden" name="comment_form_id" value="<?php bp_activity_id(); ?>" />
 					</div>
